Add files via upload
This automates the extraction of a background subtracted light curve.
You must have CIAO installed, along with its scripting tools. The folder of the galaxy must be "galaxyname_file", and its subfolders must be named by the Observation ID. Further, the "evt2.fits" file must be placed directly in the subfolder, and not in "primary" or "secondary". You must also have the "src.reg" file and "bkg.reg" files in this subfolder.