Added the Automator.ipynb Final
This automates the extraction of background subtracted light curves using DMEXTRACT. It requires CIAO to be in a conda environment, and this notebook to be run in that environment. It also requires pandas, matplotlib, astropy, and regular expression libraries to be installed. Inside the parent directory, the galaxy folder must be labeled "galaxyname_file", and all subfolders must be named with their respective Observation IDs. Finally, inside those subfolders, the "evt2.fits" file must be present directly, along with the "src.reg" and "bkg.reg" files